feat(api): further cleanup test runner api
