+++ packages/api/test_runner2.ts
@@ -1,6 +1,8 @@
 export * from './src/test_runner2/TestResult';
 export * from './src/test_runner2/TestRunner';
 export * from './src/test_runner2/TestStatus';
-export * from './src/test_runner2/RunResult';
+export * from './src/test_runner2/DryRunResult';
 export * from './src/test_runner2/RunOptions';
+export * from './src/test_runner2/MutantRunResult';
+export * from './src/test_runner2/TestSelection';
 export * from './src/test_runner2/RunStatus';
